chore: clear low-risk debt (dead loop, classify, headers, file split)
- cmd/init.go: remove the dead provider-decoration loop (its result was discarded; real decoration happens in cmd/import and cmd/export) and the now-unused internal/provider import.

- ClassifyForNautobot: route module-like types (module, adapter, transceiver) to CategoryModule instead of falling through to CategoryDevice (fixes documented bug #5).

- Add missing MIT headers to internal/provider/registry.go, internal/config/comments.go, internal/config/config_test.go, internal/core/helper.go.

- Split cmd/add/module.go (406 lines) by moving device/placement/name-resolution helpers into cmd/add/module_placement.go; both files now under the 300-line budget.
